Moved test namespace to DompdfModule namespace.
